调整发布流程功能，修复bug

+++ src/main/java/com/luna/bpm/modeler/web/controller/ModelerController.java
@@ -17,14 +17,19 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.luna.bpm.conf.service.BpmConfBaseService;
 import com.luna.bpm.process.cmd.SyncProcessCmd;
+import com.luna.bpm.process.entity.BpmProcess;
+import com.luna.bpm.process.service.BpmProcessService;
 import com.luna.common.Constants;
 import com.luna.common.utils.JsonUtil;
 
@@ -36,6 +41,12 @@ public class ModelerController {
     private ProcessEngine processEngine;
 	@Autowired
 	private SyncProcessCmd syncProcessCmd;
+	
+
+	@Autowired
+	private BpmProcessService bpmProcessService;
+	@Autowired
+	private BpmConfBaseService bpmConfBaseService;
 	
 	private final static String redirectUrl = "redirect:/modeler";
 
@@ -50,6 +61,90 @@ public class ModelerController {
         model.addAttribute("models", models);
         return "bpm/modeler/list";
     }
+    
+
+	/**
+	 * 发布流程请求
+	 * @param model
+	 * @return
+	 */
+    @RequestMapping(value="deployModel", method = RequestMethod.GET)
+    public String deployView(org.springframework.ui.Model model,
+    		@RequestParam(value = "modelId", required = false) String modelId) {
+    	BpmProcess bpmProcess = new BpmProcess();
+    	bpmProcess.setPriority(50);
+        model.addAttribute("modelId", modelId);
+        model.addAttribute("m", bpmProcess);
+        
+        return "bpm/modeler/editForm";
+    }
+    
+    /**
+     * 将模型发布为流程
+     * 
+     * @param id
+     * @return
+     * @throws Exception
+     */
+    @RequestMapping(value="deployModel", method = RequestMethod.POST)
+    public String deploy(@RequestParam("modelId") String modelId, 
+    		@ModelAttribute("m") BpmProcess bpmProcess,
+    		RedirectAttributes redirectAttributes) throws Exception {
+        RepositoryService repositoryService = processEngine.getRepositoryService();
+        Model modelData = repositoryService.getModel(modelId);
+        byte[] content = repositoryService.getModelEditorSource(modelData.getId());
+        if(content == null){
+        	redirectAttributes.addFlashAttribute(Constants.ERROR, "未找到模型，流程不能发布！");
+        	return redirectUrl;
+        }
+        JsonNode modelNode = (JsonNode) new ObjectMapper().readTree(content);
+        if (modelNode.get(EditorJsonConstants.EDITOR_CHILD_SHAPES) == null) {
+        	redirectAttributes.addFlashAttribute(Constants.ERROR, "模型中没有流程图，流程不能发布！");
+        	return redirectUrl;
+		}
+
+        BpmnModel model = new BpmnJsonConverter().convertToBpmnModel(modelNode);
+        byte[] bpmnBytes = new BpmnXMLConverter().convertToXML(model);
+
+        String processName = modelData.getName() + ".bpmn20.xml";
+        Deployment deployment = repositoryService.createDeployment()
+                .name(modelData.getName())
+                .addString(processName, new String(bpmnBytes, "UTF-8"))
+                .deploy();
+        modelData.setDeploymentId(deployment.getId());
+        repositoryService.saveModel(modelData);
+        
+        ProcessDefinition processDefinition = repositoryService.createProcessDefinitionQuery()
+                .deploymentId(deployment.getId()).singleResult();
+
+        syncProcessCmd.setProcessDefinitionId(processDefinition.getId());
+    	syncProcessCmd.setProcessDefinitionName(modelData.getName());
+        processEngine.getManagementService().executeCommand(syncProcessCmd);
+        
+        bpmProcess.setCode(processDefinition.getId());
+        bpmProcess.setBpmConfBase(bpmConfBaseService.findByProcessDefinitionId(processDefinition.getId()));
+        
+        bpmProcessService.save(bpmProcess);
+        
+        
+        /*
+        List<ProcessDefinition> processDefinitions = repositoryService
+                .createProcessDefinitionQuery()
+                .deploymentId(deployment.getId()).list();
+
+        for (ProcessDefinition processDefinition : processDefinitions) {
+        	syncProcessCmd.setProcessDefinitionId(processDefinition.getId());
+        	syncProcessCmd.setProcessDefinitionName(modelData.getName());
+            processEngine.getManagementService().executeCommand(syncProcessCmd);
+        }
+        */
+    	
+        redirectAttributes.addFlashAttribute(Constants.MESSAGE, "流程模型["+ modelData.getName() + "]已发布成["
+        				+bpmProcess.getName()+"],请在流程配置中配置相关参数！");
+        return redirectUrl;
+    }
+
+
 
     /**
      * 在modeler中打开或新建一个模型
@@ -74,8 +169,22 @@ public class ModelerController {
      * 删除一个模型
      */
     @RequestMapping("remove")
-    public String remove(@RequestParam("id") String id) {
-        processEngine.getRepositoryService().deleteModel(id);
+    public String remove(@RequestParam("id") String id,
+    		RedirectAttributes redirectAttributes) {
+    	
+    	RepositoryService repositoryService = processEngine.getRepositoryService();
+        Model model = repositoryService.createModelQuery().modelId(id).singleResult();
+        
+        
+        if(null == model.getDeploymentId() || 0 == model.getDeploymentId().length()){
+        	processEngine.getRepositoryService().deleteModel(id);
+        	redirectAttributes.addFlashAttribute(Constants.MESSAGE, "模型删除成功！");
+        }
+        else{
+        	redirectAttributes.addFlashAttribute(Constants.ERROR, "模型已发布，不能删除！");
+        }
+        
+        
         return redirectUrl;
     }
 
@@ -87,7 +196,10 @@ public class ModelerController {
      * @throws Exception
      */
     @RequestMapping("deploy")
-    public String deploy(@RequestParam("id") String id, RedirectAttributes redirectAttributes) throws Exception {
+    public String deploy(@RequestParam("id") String id, 
+    		RedirectAttributes redirectAttributes) throws Exception {
+    	
+    	
         RepositoryService repositoryService = processEngine.getRepositoryService();
         Model modelData = repositoryService.getModel(id);
         byte[] content = repositoryService.getModelEditorSource(modelData.getId());
@@ -121,6 +233,7 @@ public class ModelerController {
         	syncProcessCmd.setProcessDefinitionName(modelData.getName());
             processEngine.getManagementService().executeCommand(syncProcessCmd);
         }
+    	
         redirectAttributes.addFlashAttribute(Constants.MESSAGE, processName+" 已发布！");
         return redirectUrl;
     }
